Fix auto-updater and comments
